+++ mountpoint-s3-fs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased (v0.9.1)
 
 * Upgrade cargo dependencies.
+* Fix incorrect validation of default data cache limit which would cause 100% space usage of cache file system
 
 ## v0.9.0 (January 22, 2026)